[fix]: css 변경 및 인터페이스, api 통신 수정
